requirements.txt
pyyaml
wday
tkcode
tkterminal
confscript
ttkbootstrap
requests
terminal-animation